Enforce `RaidLevel` member count bounds in metadata and assembly:
- Centralized minimum and maximum member count definitions under `RaidLevel`.
- Validated member count in `ArraySuperblock::decode`, rejecting unbuildable configurations for RAID5 (<3 members) and RAID6 (<4 or >257 members).
- Updated RAID5/RAID6 engines to use shared bounds, ensuring consistency across metadata and runtime logic.
- Added tests verifying boundary conditions for member count validation.
- Updated documentation to reflect member count constraints and shared `RaidLevel` definitions.

+++ drivers/storage/raid/src/gf256.rs
@@ -62,7 +62,7 @@ pub(crate) const fn gpow(exp: u64) -> u8 {
 /// The largest data-member count a GF(2^8) Q syndrome keeps distinct,
 /// non-zero coefficients for: `g⁰ … g²⁵⁴`. A double-parity array may not have
 /// more than this many data members.
-pub(crate) const MAX_DATA_MEMBERS: u64 = 255;
+pub(crate) const MAX_DATA_MEMBERS: u16 = 255;
 
 /// The multiplicative inverse of `x` in GF(2^8). `x` must be non-zero (zero has
 /// no inverse); the caller guarantees this. Computed as `x^254` (Fermat:
